+++ apps/docs/src/examples/select/preview.tsx
@@ -7,6 +7,11 @@ export default function Example() {
 			<Select.Item>Qwik</Select.Item>
 			<Select.Item>Vue</Select.Item>
 			<Select.Item>Svelte</Select.Item>
+			<Select.Section>
+				<Select.Label>Others</Select.Label>
+				<Select.Item>Angular</Select.Item>
+				<Select.Item>JQuery</Select.Item>
+			</Select.Section>
 		</Select>
 	)
 }