+++ packages/styles/src/components/tooltip/variants.ts
@@ -3,7 +3,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { tv } from "tailwind-variants"
+import { popoverBase } from "../../utils"
 
 export const tooltipVariants = tv({
-	base: "m-1 max-w-sm rounded-md border border-border bg-background p-2 text-foreground shadow-xl outline-none [&[data-entering]]:animate-fade [&[data-exiting]]:animate-fadeOut",
+	base: ["p-md", popoverBase()],
 })