+++ packages/ui/src/components/ui/button/button.tsx
@@ -5,12 +5,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import type React from "react"
-import { forwardRef } from "react"
 import type { VariantProps } from "tailwind-variants"
 
 import { buttonVariants } from "@pixelshades/styles/components/button"
 
-import { RenderSlot } from "@pixelshades/utils/jsx"
+import { RenderSlot, forwardRef } from "@pixelshades/utils/jsx"
 import {
 	Button as AriaButton,
 	type ButtonProps as AriaButtonProps,
@@ -23,30 +22,14 @@ import { LoadingSpinner } from "../loading-spinner"
 type ButtonVariantProps = VariantProps<typeof buttonVariants>
 
 interface ButtonProps extends ButtonVariantProps, AriaButtonProps {
+	className?: string
 	isLoading?: boolean
-	onClick?: AriaButtonProps["onPress"]
 	before?: React.ReactElement<HTMLElement>
 	after?: React.ReactElement<HTMLElement>
 }
 
-const Button = forwardRef<HTMLButtonElement, ButtonProps>(
-	(
-		{
-			size,
-			variant,
-			className,
-			children,
-			isLoading,
-			onClick,
-			onPress,
-			isDisabled,
-			before,
-			after,
-			withRing,
-			...props
-		},
-		ref,
-	) => {
+const Button = forwardRef(
+	({ size, variant, className, children, isLoading, isDisabled, before, after, withRing, ...props }: ButtonProps) => {
 		const buttonGroupState = buttonGroupContext.useStyleContext()
 
 		/* If Button is in Button Group apply Button Group Styles, this can still be overwritten on button layer **/
@@ -57,27 +40,18 @@ const Button = forwardRef<HTMLButtonElement, ButtonProps>(
 						withRing,
 						...values,
 						size,
-						className: typeof className === "function" ? className(values) : className,
+						className,
 					})
 				: buttonVariants({
 						variant,
 						withRing,
 						...values,
 						size,
-						className: typeof className === "function" ? className(values) : className,
+						className,
 					})
 
 		return (
-			<AriaButton
-				className={buttonStyles}
-				onPress={(e) => {
-					onPress?.(e)
-					onClick?.(e)
-				}}
-				isDisabled={isDisabled || isLoading}
-				ref={ref}
-				{...props}
-			>
+			<AriaButton className={buttonStyles} isDisabled={isDisabled || isLoading} {...props}>
 				<>
 					{isLoading && <LoadingSpinner className="size-4 text-inherit" />}
 					<If condition={before && !isLoading}>