+++ packages/ui/src/components/ui/menu/menu.tsx
@@ -7,7 +7,6 @@
 import { menuVariants } from "@pixelshades/styles/components/menu"
 import type { HTMLAttributes } from "react"
 
-import { forwardRef } from "@pixelshades/utils/jsx"
 import { createStyleContext } from "@pixelshades/utils/styles"
 import { ChevronRight } from "lucide-react"
 import {
@@ -18,32 +17,37 @@ import {
 	type MenuProps as AriaMenuProps,
 	MenuTrigger as AriaMenuTrigger,
 	type MenuTriggerProps as AriaMenuTriggerProps,
-	Popover as AriaPopover,
 	Section as AriaSection,
 	SubmenuTrigger as AriaSubMenuTrigger,
+	type PopoverProps,
 } from "react-aria-components"
 import { Kbd } from "../kbd"
+import { Popover } from "../popover"
 import { Separator } from "../separator"
 
 const { withContext, withProvider } = createStyleContext(menuVariants)
 
-const { menuPopover, content, shortcut } = menuVariants()
+const { content, popover } = menuVariants()
 
 export type MenuProps = AriaMenuTriggerProps
+
+export type MenuContentProps<T extends object> = {
+	placement?: PopoverProps["placement"]
+} & AriaMenuProps<T>
+
+const MenuContent = <T extends object>({ children, className, placement, ...props }: MenuContentProps<T>) => (
+	<Popover className={popover({ className })} placement={placement}>
+		<AriaMenu {...props} className={content({ className })}>
+			{children}
+		</AriaMenu>
+	</Popover>
+)
+
+MenuContent.displayName = "MenuContent"
 
 const MenuRoot = withProvider(AriaMenuTrigger)
 
 const MenuSection = AriaSection
-
-export type MenuContentProps<T extends object> = AriaMenuProps<T>
-
-const MenuContent = forwardRef(<T extends object>({ children, className, ...props }: MenuContentProps<T>) => (
-	<AriaPopover isNonModal {...props} className={menuPopover()}>
-		<AriaMenu {...props} className={content({ className })}>
-			{children}
-		</AriaMenu>
-	</AriaPopover>
-))
 
 export type MenuItemProps = AriaMenuItemProps & { className?: string }
 
@@ -86,3 +90,5 @@ export const Menu = Object.assign(MenuRoot, {
 	SubMenu: MenuSubMenu,
 	SubMenuTrigger: MenuSubMenuTrigger,
 })
+
+export { MenuContent, MenuSection, MenuItem, MenuLabel, MenuShortcut, MenuSeparator, MenuSubMenu, MenuSubMenuTrigger }